src/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.127.0",
    "google-genai>=1.56.0",
    "supabase>=2.27.0",
    "uvicorn>=0.40.0",
]
